✨ Refactor API imports to use championApi and leaderboardApi; remove serverApi
